# Quick Optimization: Memoize prop-name conversions for IdsElement attribChangeCallback