**Explain the details for making this change. What existing problem does the pull request solve?**
<!--
Example: When "Adding a function to do X",
explain why it is necessary to have a way to do X.
-->
When setting any prop in IdsElement, we were going through `number-of-chars` ^ 2 or 3 before being able to run through normal setter logic as we were converting `attrib-name` to `attribName` for property assignment. 

Through some benchmarking, found that using memoization could get that down a bit. In some sample test code (shared in MSTeams), running a similar operation gave these results on my work PC:

```
ITERATIONS: 10
withoutMemo: 0.207ms
withMemo: 0.037ms
change: ~540%

For 100 iterations:

ITERATIONS: 100
withoutMemo: 0.331ms
withMemo: 0.057ms
change: ~580%

ITERATIONS: 3000
withoutMemo: 4.322ms
withMemo: 0.226ms
change: ~1900% (!)
```

Note that these tests are much more simple and repeating one property, real world scenario should be better.

When loading `http://localhost:4300`, only on initial render we had *3138 cache hits* (times that we could re-load the var without running `camelCase` op), and for larger production apps this will likely be larger. So we are saving some time in interactions or load up (especially on slower PCs or with multitasking) which may be useful to help us maintain easy 60fps or not worry about less critical things.

**Related github/jira issue (required)**:
<!--
Provide a link to the related issue(s) to this Pull Request;
auto-closing github issues if necessary (example: "Closes #100" or "Fixes #1230")
-->
Not listed on #140  but related. Most similar is to how we can optimize style loading on each component instance.

**Steps necessary to review your pull request (required)**:
<!--
Include:
- commands you ran and their output
- screenshots / videos
- test scenarios
-->
- N/A, app should run as it normally does and just need to confirm tests pass. Change is minor/mainly for perf with real app usage.

**Included in this Pull Request**:
- [ ] An e2e or functional test for the bug or feature.
- [ ] A note to the change log.

<!-- After submitting your PR, please check back to make sure checks on the PR -->